Updates Makefile for exams
Removes DELETE_ON_ERROR, we don't want to lose all our progress in case
we put a lot of effort in using interactive mode and get an error.

If we add all questions as databases, we will regenerate older exams
when we build new ones --- since only the newest exam has no
dependencies, we will start with that one and generate them in inverse
chronological order.