Changes repo structure
The `modules/` directory no longer houses only submodules. The
`modules/` directory is the home of each learning module in the course,
which in turn houses subdmodules.

Thus the learning-platform stuff now moves to the respective module in
`modules/`.

The next step is to make the modules the "interface" for `release/` and
`studyguide`, so that they don't have to reach deep inside the
submodules inside the modules --- they should only care about the
modules, which in turn keep track of the submodules they use and how.